Split classes and utility functions into separate files
